{/src/main/java/com/oaknorth/drools/RatioValue.java}

+++ src/main/java/com/oaknorth/drools/RatioValue.java
@@ -21,7 +21,7 @@ public class RatioValue implements java.io.Serializable {
 	private double data;
 
 	@org.kie.api.definition.type.Label("Absolute Anomaly")
-	private double absoluteAnomaly;
+	private int absoluteAnomaly;
 
 	@org.kie.api.definition.type.Label("Ratio Name")
 	private java.lang.String ratioName;
@@ -61,16 +61,16 @@ public class RatioValue implements java.io.Serializable {
 		this.ratioName = ratioName;
 	}
 
-	public double getAbsoluteAnomaly() {
+	public int getAbsoluteAnomaly() {
 		return this.absoluteAnomaly;
 	}
 
-	public void setAbsoluteAnomaly(double absoluteAnomaly) {
+	public void setAbsoluteAnomaly(int absoluteAnomaly) {
 		this.absoluteAnomaly = absoluteAnomaly;
 	}
 
 	public RatioValue(java.lang.Long id, java.lang.String financialYear,
-			double data, double absoluteAnomaly, java.lang.String ratioName) {
+			double data, int absoluteAnomaly, java.lang.String ratioName) {
 		this.id = id;
 		this.financialYear = financialYear;
 		this.data = data;
